backend/requirements.txt
# Framework principal
fastapi>=0.104.1
uvicorn[standard]>=0.24.0

# Base de datos
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23
alembic>=1.13.1

# Autenticación y seguridad
passlib[bcrypt]>=1.7.4
python-jose>=3.3.0
bcrypt>=4.1.2
cryptography>=41.0.8
hvac>=2.0.0  # HashiCorp Vault client
bleach>=6.2.0  # HTML sanitization para XSS protection

# Utilidades
python-multipart>=0.0.6
pydantic-settings>=2.1.0

# Caché
redis[hiredis]>=5.0.1

# Logging
python-json-logger>=2.0.7

# Testing
pytest>=7.4.3
httpx>=0.25.2
pytest-asyncio>=0.21.1

# Producción adicional
gunicorn>=21.2.0
